+++ src/scanner.rs
@@ -51,7 +51,7 @@ impl Scanner {
 
     /// Return the range of the entire document
     pub fn document_range(&self) -> NodeRange {
-        ((0, 0), (self.line as u16, self.col as u16))
+        ((0, 0), (self.line as u32, self.col as u32))
     }
 
     /// Scans the source file into a token stream `tokens` and returns a result containing a
@@ -853,14 +853,14 @@ impl Scanner {
             return;
         }
         self.tokens
-            .push(Token::new(t, self.line as u16, self.start_of_token as u16));
+            .push(Token::new(t, self.line as u32, self.start_of_token as u32));
     }
 
     fn push_named_token(&mut self, t: TokenType, name: &str) {
         self.tokens.push(Token::named(
             t,
-            self.line as u16,
-            self.start_of_token as u16,
+            self.line as u32,
+            self.start_of_token as u32,
             name,
         ));
     }
